New class for the Maxbotix ultrasonic distance sensor
The WPI Ultrasonic class is geared to the Daventech SRF04 sensor, which
has a 5.8us/mm pulse output.  The Maxbotix sensor that came in the kit
of parts in 2018 has a 1us/mm pulse output.  This class derives from the
WPI Ultrasonic class but re-implements the GetRangeInches and GetRangeMM
functions to use the proper conversion factor.